Restructure the frontend to align with the core
Bring the core game code into src/core. This is simply a copy of
the main SGT Git repo, so in future we could pull this in from
source to automate updates. Add the necessary core source files
to the Makefile.

Rename game_collection.* to frontend.*, as this will provide the
frontend expected by the midend. Add the necessary stub functions
to frontend.c to satisy the linker.

Remove all use of flex and heap, because the core is tightly wedded
to malloc. For simplicity, it is better to standardise on malloc
and take the hit on 'losing' memory.