Add support for mouse dragging
Implement mouse dragging, restructuring the clicks to operate on
release so that we can differentiate between a drag and a click
before we issue a _RELEASE event to the mid-end.

MIDDLE_ clicks and drags are emulated using Ctrl-Select, to avoid
any confusion with the Menu button.

Change numeric keypad handling to use osbyte_IN_KEY for consistency.