Begin to add blitter support to front end
Begin to add the basic structure for blitter support to the front-end,
with the ability to create dummy blitter objects on request from the
mid-end, and destroy them when a front-end is closed.

A commit of work in progress, to allow more restructuring of sprite
support in order to reduce code duplication when creating and updating
sprite areas.