Begin work on support for Configuration
Add a generic Game Config dialogue to take lists of config_items and
turn them into dialogue boxes. This covers Custom... game types,
Preferences..., Specific game... and Random seed...

Support is currently present for text and option fields, but no attempt
is made to process combo boxes. The returning of text fields is less
than optimal, with a free() and strdup() for each one regardless of
whether the value has changed or would fit into the existing allocation.

No Interactive Help is included.

There is no Save option for Preferences.

Changing Preferences causes a new game to be generated, even if this
wasn't what the user expected.

The sizing of fields is odd, with the width of the template label icon
being included in the field width to the right of the "left" line.

There is still debug output present in the modules.