Remove horizontal scroll
The index window has no concept of scrolling horizontally, so remove the
ability to do so from the scroll event handler.

This removes the possibility of incorrectly adjusting the xscroll
offset, which could cause the icons to shift horizontally. This may be
simply masking another issue with the window extents!