Handle early status bar updates
Allow the text in the status bar to be updated from the moment that the
game window has been created, to catch updates made by the backend
before the window has been opened.

This means that we must allocate the buffer within the game_window_block
struct instead of dynamically, because we don't know until the window
is opened whether the backend will request the bar or not.