Fix redraw glitches when game window is resized
Call midend_force_redraw() instead of midend_redraw() in
frontend_negotiate_game_size(), so that the canvas is redrawn completely
after the change of dimensions.

Note that it isn't entirely clear why this works. The problematic calls
to frontend_negotiate_game_size() are occurring immediately after calls
to midend_new_game(), and the API documentation states that

    after calling this function and after calling midend_size(), you
    should then call midend_redraw(). (It is not necessary to call
    midend_force_redraw(); that will discard the draw state and create a
    fresh one, which is unnecessary in this case since there's a fresh
    one already. It would work, but it's usually excessive.)

It does appear to clear the problem, though, and the documentation
does indicate that midend_force_redraw() will work.