Include contributors in Program Info dialogue
Add an "Additional Contributors" field to the Program Info dialogue, and
arrange for this to be populated with the names credited in the upstream
manual.

This is currently a manual process, requiring the
src/additional_contributors.h file to be updated if the manual changes.
The text also contains un-translated English text alongside the names,
so an alternative approach may be required if internationalisation is
applied in the future.