Implement the text_fallback() function
Implement the text_fallback() function to allow games to access
characters from the top half of the current encoding via UTF-8
conversion.

Requested strings are looked up using the current UCS conversion table
obtained from the OS. This allows the small collection of characters
which games currently need (the mathematical symbols for minus, times
and divide) to be displayed without recourse to a full Unicode
implementation.

The functionality relies on Service_International 8, and so is not
available on OS versions which do not support this. In this case, text
should continue to be displayed in 7-bit ASCII as previously.